fixing

+++ backend/api/views.py
@@ -5,27 +5,23 @@ from django.http import HttpResponse
 from django.shortcuts import get_object_or_404
 from django_filters.rest_framework import DjangoFilterBackend
 from djoser.views import UserViewSet
+from recipes.models import (Favorite, Ingredient, Recipe, RecipeIngredient,
+                            ShoppingCart, Tag)
 from rest_framework import permissions, status, viewsets
 from rest_framework.decorators import action
 from rest_framework.exceptions import ValidationError
 from rest_framework.permissions import AllowAny
 from rest_framework.response import Response
+from users.models import User, Subscription
 
-from api.filters import IngredientFilter, RecipeFilter
-from api.permissions import AuthorOrReadOnly
-from api.serializers import (CreateRecipeSerializer, FavoriteSerializer,
-                             IngredientSerializer, RecipeSerializer,
-                             ShoppingCartSerializer,
-                             SubscriptionPresentSerializer,
-                             SubscriptionSerializer, TagSerializer,
-                             UserSerializer)
-from recipes.models import (Favorite,
-                            Ingredient,
-                            Recipe,
-                            RecipeIngredient,
-                            ShoppingCart,
-                            Tag)
-from users.models import Subscription, User
+from .filters import IngredientFilter, RecipeFilter
+from .permissions import AuthorOrAdminOrReadOnly
+from .serializers import (CreateRecipeSerializer, UserSerializer,
+                          FavoriteSerializer, IngredientSerializer,
+                          RecipeSerializer, ShoppingCartSerializer,
+                          SubscriptionSerializer,
+                          SubscriptionPresentSerializer,
+                          TagSerializer)
 
 
 class UserViewSet(UserViewSet):
@@ -50,28 +46,24 @@ class UserViewSet(UserViewSet):
 
     @action(
         detail=True,
-        methods=['post'],
+        methods=['post', 'delete'],
         url_path='subscribe',
         url_name='subscribe',
     )
     def to_subscribe(self, request, id):
         author = get_object_or_404(User, id=id)
         if request.method == 'POST':
-            serializer = self.get_serializer(data={
-                'subscriber': request.user.id,
-                'author': author.id})
+            data = {'subscriber': request.user.id, 'author': author.id}
+            serializer = self.get_serializer(data=data)
             serializer.is_valid(raise_exception=True)
             serializer.save()
             return Response(
                 serializer.data, status=status.HTTP_201_CREATED
             )
-
-    @to_subscribe.mapping.delete
-    def delete_subscription(self, request, id):
-        author = get_object_or_404(User, id=id)
-        subscription = get_object_or_404(Subscription.objects.filter(
-                                         subscriber=request.user,
-                                         author=author))
+        subscription = Subscription.objects.filter(subscriber=request.user,
+                                                   author=author)
+        if not subscription.exists():
+            raise ValidationError('Такой подписки нет')
         subscription.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
 
@@ -83,11 +75,23 @@ class UserViewSet(UserViewSet):
     )
     def get_subscription_list(self, request):
         authors = User.objects.filter(author__subscriber=request.user)
-        return self.get_paginated_response(self.get_serializer(
-            self.paginate_queryset(queryset=authors), many=True).data)
+        result_pages = self.paginate_queryset(
+            queryset=authors
+        )
+        serializer = self.get_serializer(result_pages, many=True)
+        return self.get_paginated_response(serializer.data)
+
+
+class TagViewSet(viewsets.ReadOnlyModelViewSet):
+
+    queryset = Tag.objects.all()
+    serializer_class = TagSerializer
+    pagination_class = None
+    permission_classes = (AllowAny, )
 
 
 class IngredientViewSet(viewsets.ReadOnlyModelViewSet):
+
     queryset = Ingredient.objects.all()
     serializer_class = IngredientSerializer
     filter_backends = (DjangoFilterBackend,)
@@ -97,17 +101,10 @@ class IngredientViewSet(viewsets.ReadOnlyModelViewSet):
     permission_classes = (AllowAny, )
 
 
-class TagViewSet(viewsets.ReadOnlyModelViewSet):
-    queryset = Tag.objects.all()
-    serializer_class = TagSerializer
-    pagination_class = None
-    permission_classes = (AllowAny, )
-
-
 class RecipeViewSet(viewsets.ModelViewSet):
     serializer_class = RecipeSerializer
     permission_classes = (
-        permissions.IsAuthenticatedOrReadOnly, AuthorOrReadOnly
+        permissions.IsAuthenticatedOrReadOnly, AuthorOrAdminOrReadOnly
     )
     filter_backends = (DjangoFilterBackend,)
     filterset_class = RecipeFilter
@@ -152,10 +149,11 @@ class RecipeViewSet(viewsets.ModelViewSet):
             return Response(
                 serializer.data, status=status.HTTP_201_CREATED
             )
+        model = ShoppingCartSerializer.Meta.model
         if self.request.user.is_authenticated:
-            queryset = ShoppingCartSerializer.Meta.model.objects.filter(
-                user=request.user,
-                recipe=get_object_or_404(Recipe, pk=pk))
+            queryset = model.objects.filter(user=request.user,
+                                            recipe=get_object_or_404(
+                                                Recipe, pk=pk))
             if not queryset.exists():
                 return Response(status=status.HTTP_400_BAD_REQUEST)
             queryset.delete()
@@ -178,10 +176,11 @@ class RecipeViewSet(viewsets.ModelViewSet):
             return Response(
                 serializer.data, status=status.HTTP_201_CREATED
             )
+        model = FavoriteSerializer.Meta.model
         if self.request.user.is_authenticated:
-            queryset = FavoriteSerializer.Meta.model.objects.filter(
-                user=request.user,
-                recipe=get_object_or_404(Recipe, pk=pk))
+            queryset = model.objects.filter(user=request.user,
+                                            recipe=get_object_or_404(
+                                                Recipe, pk=pk))
             if not queryset.exists():
                 return Response(status=status.HTTP_400_BAD_REQUEST)
             queryset.delete()
@@ -196,7 +195,7 @@ class RecipeViewSet(viewsets.ModelViewSet):
         url_name='download_shopping_cart',
     )
     def download_shopping_cart(self, request):
-        queryset = (
+        ingredients_cart = (
             RecipeIngredient.objects.filter(
                 recipe__shopping_cart__user=request.user
             ).values(
@@ -205,14 +204,14 @@ class RecipeViewSet(viewsets.ModelViewSet):
             ).annotate(ingredient_value=Sum('amount')
                        ).order_by('ingredient__name')
         )
-        if not queryset.exists():
+        if not ingredients_cart.exists():
             raise ValidationError('В списке покупок нет добавленных рецептов')
         response = HttpResponse(
-            content_type='text/csv',
+            content_type="text/csv",
             headers={'Content-Disposition':
                      'attachment;filename="shopping_cart.csv"'},)
         writer = csv.DictWriter(response,
-                                fieldnames=queryset.first().keys())
+                                fieldnames=ingredients_cart.first().keys())
         writer.writeheader()
-        writer.writerows(queryset)
+        writer.writerows(ingredients_cart)
         return response
